Changed the way to compute the hash for playlist to take in account any change change that will happens

+++ src/libraries/os/beatSaber/installer/PlaylistInstaller.ts
@@ -33,7 +33,10 @@ export default class PlaylistInstaller {
   }
 
   public static async Uninstall(playlist: PlaylistLocal): Promise<void> {
-    await fs.unlink(playlist.path);
+    if (playlist.path) {
+      await fs.unlink(playlist.path);
+    }
+
     PlaylistLibrary.RemovePlaylist(playlist);
   }
 }
